+++ lib/akasha/async_event_router.rb
@@ -9,13 +9,16 @@ module Akasha
     DEFAULT_PAGE_SIZE = 20
     DEFAULT_PROJECTION_STREAM = 'AsyncEventRouter'.freeze
     DEFAULT_CHECKPOINT_STRATEGY = Akasha::Checkpoint::HttpEventStoreCheckpoint
+    STREAM_NAME_SEP = '-'.freeze
 
-    def connect!(repository, projection_name: DEFAULT_PROJECTION_STREAM,
+    def connect!(repository, projection_name: nil,
                  checkpoint_strategy: DEFAULT_CHECKPOINT_STRATEGY,
                  page_size: DEFAULT_PAGE_SIZE, poll: DEFAULT_POLL_SECONDS)
+      projection_name = projection_name(repository) if projection_name.nil?
       projection_stream = repository.store.streams[projection_name]
       checkpoint = checkpoint_strategy.is_a?(Class) ? checkpoint_strategy.new(projection_stream) : checkpoint_strategy
-      repository.merge_all_by_event(into: projection_name, only: registered_event_names)
+      repository.merge_all_by_event(into: projection_name,
+                                    only: registered_event_names)
       Thread.new do
         run_forever(projection_stream, checkpoint, page_size, poll)
       end
@@ -39,6 +42,13 @@ module Akasha
           end
         end
       end
+    end
+
+    def projection_name(repository)
+      parts = []
+      parts << repository.namespace unless repository.namespace.nil?
+      parts << DEFAULT_PROJECTION_STREAM
+      parts.join(STREAM_NAME_SEP)
     end
   end
 end