+++ spec/akasha/async_event_router_spec.rb
@@ -1,58 +1,71 @@
 
 # TODO: Simplify initialization!
 describe Akasha::AsyncEventRouter, integration: true do
-  let(:repository) { Akasha::Repository.new(store) }
+  let(:repository) { Akasha::Repository.new(store, namespace: unique_namespace) }
+  let(:unique_namespace) { SecureRandom.uuid }
   let(:store) { Akasha::Storage::HttpEventStore.new(http_es_config) }
-  let(:projection_name) { gensym(:projection) }
-  let(:stream_name) { gensym(:stream) }
-  let(:name_changed_event) { gensym(:name_changed) }
-  let(:something_happened_event) { gensym(:something_happened) }
-  let(:ignored_event) { gensym(:ignored_event) }
-
-  let(:events) do
-    [
-      Akasha::Event.new(name_changed_event, old_name: nil, new_name: 'new name'),
-      Akasha::Event.new(something_happened_event),
-      Akasha::Event.new(name_changed_event, old_name: 'new_name', new_name: 'newest name'),
-      Akasha::Event.new(ignored_event)
-    ]
-  end
+  let(:item) { Item.new('item-1') }
 
   before do
     Akasha::Aggregate.connect!(repository)
-    subject.register_event_listener(name_changed_event, listener)
-    subject.register_event_listener(something_happened_event, listener)
-    @thread = subject.connect!(repository, projection_name: projection_name)
+    subject.register_event_listener(:name_changed, listener)
+    subject.register_event_listener(:count_changed, listener)
+    @thread = subject.connect!(repository)
   end
 
   after do
     @thread.kill
   end
 
-  context 'listener handling name_changed and something_happened' do
+  context 'listener handling name_changed and count_changed' do
     let(:listener) { FakeEventListener.new }
 
+    before do
+      item.name = 'new name'
+      item.count = 5
+      item.name = 'newest name'
+      item.save!
+    end
+
     it 'routes events to listeners' do
-      store.streams[stream_name].write_events(events)
       wait(10).for { listener.calls.size }.to eq 3
     end
 
     it 'preserves ordering' do
-      store.streams[stream_name].write_events(events)
-      expected_calls = [
-        :"on_#{name_changed_event}",
-        :"on_#{something_happened_event}",
-        :"on_#{name_changed_event}"
+      expected_calls = %i[
+        on_name_changed
+        on_count_changed
+        on_name_changed
       ]
       wait(10).for { listener.calls }.to eq expected_calls
     end
+
+    context 'for events within another namespace' do
+      let(:another_namespace) { SecureRandom.uuid }
+      let(:another_repo) { Akasha::Repository.new(store, namespace: another_namespace) }
+
+      it 'ignores those events' do
+        Akasha::Aggregate.connect!(another_repo)
+        item.name = 'different name'
+        item.count = 555
+        item.save!
+
+        wait(10).for { listener.calls.size }.to eq 3
+      end
+    end
   end
 
-  context 'listener failing for something_happened' do
-    let(:listener) { FakeEventListener.new(fail_on: [something_happened_event]) }
+  context 'listener failing for count_changed' do
+    let(:listener) { FakeEventListener.new(fail_on: [:count_changed]) }
+
+    before do
+      item.name = 'new name'
+      item.count = 5
+      item.name = 'newest name'
+      item.save!
+    end
 
     it 'handles the remaining two events' do
-      store.streams[stream_name].write_events(events)
       wait(10).for { listener.calls.size }.to eq 2
     end
   end