+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Version xx
+
+* Optional namespacing for aggregate/projection streams and events allowing for isolation
+  between applications. [#12](https://github.com/bilus/akasha/pull/12)
+
+
 ## Version 0.3.0
 
 * Asynchronous event listeners (`AsyncEventRouter`). [#9](https://github.com/bilus/akasha/pull/9)